move auth-interceptor to `cores/` folder
